backend/requirements.txt
flask
flask-pymongo
flask-cors
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
